Thin out test fixtures
