+++ packages/runtime/spec/support/msg-longs-fixture.ts
@@ -1,0 +1,107 @@
+
+/**
+ * This data is generated by serializin the `spec.LongsMessage`
+ * using Grpc.AspNetCore v2.30.0 (C#).
+ *
+ * Each field has the appropriate minimum or maximum
+ * value of the native type.
+ */
+export const msg_longs_bytes = new Uint8Array([17, 255, 255, 255, 255, 255, 255, 255, 255, 24, 128, 128, 128, 128, 128, 128, 128, 128, 128, 1, 32, 255, 255, 255, 255, 255, 255, 255, 255, 127, 41, 0, 0, 0, 0, 0, 0, 0, 128, 49, 255, 255, 255, 255, 255, 255, 255, 127, 56, 255, 255, 255, 255, 255, 255, 255, 255, 255, 1, 64, 254, 255, 255, 255, 255, 255, 255, 255, 255, 1, 80, 255, 255, 255, 255, 255, 255, 255, 255, 255, 1, 97, 255, 255, 255, 255, 255, 255, 255, 255, 104, 128, 128, 128, 128, 128, 128, 128, 128, 128, 1, 112, 255, 255, 255, 255, 255, 255, 255, 255, 127, 121, 0, 0, 0, 0, 0, 0, 0, 128, 129, 1, 255, 255, 255, 255, 255, 255, 255, 127, 136, 1, 255, 255, 255, 255, 255, 255, 255, 255, 255, 1, 144, 1, 254, 255, 255, 255, 255, 255, 255, 255, 255, 1, 160, 1, 255, 255, 255, 255, 255, 255, 255, 255, 255, 1, 177, 1, 255, 255, 255, 255, 255, 255, 255, 255, 184, 1, 128, 128, 128, 128, 128, 128, 128, 128, 128, 1, 192, 1, 255, 255, 255, 255, 255, 255, 255, 255, 127, 201, 1, 0, 0, 0, 0, 0, 0, 0, 128, 209, 1, 255, 255, 255, 255, 255, 255, 255, 127, 216, 1, 255, 255, 255, 255, 255, 255, 255, 255, 255, 1, 224, 1, 254, 255, 255, 255, 255, 255, 255, 255, 255, 1, 240, 1, 255, 255, 255, 255, 255, 255, 255, 255, 255, 1]);
+
+
+/**
+ * The file `msg-long.bin`, read by the Javascript package
+ * google-protobuf v3.12.2.
+ */
+export const msg_longs_deserialized_by_google_protobuf = {
+    "fixed64FieldMin": 0,
+    "fixed64FieldMax": 18446744073709552000,
+    "int64FieldMin": -9223372036854776000,
+    "int64FieldMax": 9223372036854776000,
+    "sfixed64FieldMin": -9223372036854776000,
+    "sfixed64FieldMax": 9223372036854776000,
+    "sint64FieldMin": -9223372036854776000,
+    "sint64FieldMax": 9223372036854776000,
+    "uint64FieldMin": 0,
+    "uint64FieldMax": 18446744073709552000,
+    "fixed64FieldMinStr": "0",
+    "fixed64FieldMaxStr": "18446744073709551615",
+    "int64FieldMinStr": "-9223372036854775808",
+    "int64FieldMaxStr": "9223372036854775807",
+    "sfixed64FieldMinStr": "-9223372036854775808",
+    "sfixed64FieldMaxStr": "9223372036854775807",
+    "sint64FieldMinStr": "-9223372036854775808",
+    "sint64FieldMaxStr": "9223372036854775807",
+    "uint64FieldMinStr": "0",
+    "uint64FieldMaxStr": "18446744073709551615",
+    "fixed64FieldMinNum": 0,
+    "fixed64FieldMaxNum": 18446744073709552000,
+    "int64FieldMinNum": -9223372036854776000,
+    "int64FieldMaxNum": 9223372036854776000,
+    "sfixed64FieldMinNum": -9223372036854776000,
+    "sfixed64FieldMaxNum": 9223372036854776000,
+    "sint64FieldMinNum": -9223372036854776000,
+    "sint64FieldMaxNum": 9223372036854776000,
+    "uint64FieldMinNum": 0,
+    "uint64FieldMaxNum": 18446744073709552000
+};
+
+// msg-longs.proto:
+//
+// syntax = "proto3";
+// package spec;
+//
+//
+// message LongsMessage {
+//
+//     fixed64 fixed64_field_min = 1 [jstype = JS_NORMAL];
+//     fixed64 fixed64_field_max = 2 [jstype = JS_NORMAL];
+//
+//     int64 int64_field_min = 3 [jstype = JS_NORMAL];
+//     int64 int64_field_max = 4 [jstype = JS_NORMAL];
+//
+//     sfixed64 sfixed64_field_min = 5;
+//     sfixed64 sfixed64_field_max = 6;
+//
+//     sint64 sint64_field_min = 7;
+//     sint64 sint64_field_max = 8;
+//
+//     uint64 uint64_field_min = 9;
+//     uint64 uint64_field_max = 10;
+//
+//     //
+//
+//     fixed64 fixed64_field_min_str = 11 [jstype = JS_STRING];
+//     fixed64 fixed64_field_max_str = 12 [jstype = JS_STRING];
+//
+//     int64 int64_field_min_str = 13 [jstype = JS_STRING];
+//     int64 int64_field_max_str = 14 [jstype = JS_STRING];
+//
+//     sfixed64 sfixed64_field_min_str = 15 [jstype = JS_STRING];
+//     sfixed64 sfixed64_field_max_str = 16 [jstype = JS_STRING];
+//
+//     sint64 sint64_field_min_str = 17 [jstype = JS_STRING];
+//     sint64 sint64_field_max_str = 18 [jstype = JS_STRING];
+//
+//     uint64 uint64_field_min_str = 19 [jstype = JS_STRING];
+//     uint64 uint64_field_max_str = 20 [jstype = JS_STRING];
+//
+//     //
+//
+//     fixed64 fixed64_field_min_num = 21 [jstype = JS_NUMBER];
+//     fixed64 fixed64_field_max_num = 22 [jstype = JS_NUMBER];
+//
+//     int64 int64_field_min_num = 23 [jstype = JS_NUMBER];
+//     int64 int64_field_max_num = 24 [jstype = JS_NUMBER];
+//
+//     sfixed64 sfixed64_field_min_num = 25[jstype = JS_NUMBER];
+//     sfixed64 sfixed64_field_max_num = 26 [jstype = JS_NUMBER];
+//
+//     sint64 sint64_field_min_num = 27 [jstype = JS_NUMBER];
+//     sint64 sint64_field_max_num = 28 [jstype = JS_NUMBER];
+//
+//     uint64 uint64_field_min_num = 29 [jstype = JS_NUMBER];
+//     uint64 uint64_field_max_num = 30 [jstype = JS_NUMBER];
+//
+// }
+
